+++ packages/terra-table/CHANGELOG.md
@@ -6,8 +6,17 @@
   * Renamed `hasColumnHeaders` prop to `hasVisibleColumnHeaders`.
 
 * Added
+  * Added title attribute to column header cells to enable tooltips.
+
+* Fixed
+  * Fixed tab index issues in the column header cells.
+  
+## 5.1.1-alpha.1 - (November 9, 2023)
+
+* Added
   * Added row selection mode to the table component.
   * Added ability for consumers to specify a custom screen reader message for masked cells.
+  * Added sections support to the table component.
 
 * Fixed
   * Fixed the column header background color for the Orion Fusion theme.